requirements.txt
# Core Django
Django
djangorestframework
django-cors-headers
django-filter
django-extensions

# Database
psycopg2-binary
dj-database-url

# Task Queue
celery
redis
django-celery-beat
django-celery-results

# File Processing
pandas
openpyxl
xlrd

# AI & ML
openai
openai-agents

# Frontend
streamlit
plotly

# Monitoring & Logging
sentry-sdk

# Development
python-dotenv
django-debug-toolbar
requests

# Streamlit Frontend
streamlit
pandas
plotly
altair

# Testing
pytest
pytest-django
factory-boy

# Production
gunicorn
whitenoise
